# internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DockMind API",
    "version": "1.0.0",
    "description": "DockMind is a lightweight daemon that manages the lifecycle of an AI inference server running on an external GPU."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/status": {
      "get": {
        "summary": "Get current system state and component health",
        "responses": {
          "200": {
            "description": "Current system state and component health",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/power/on": {
      "post": {
        "summary": "Power on the eGPU and start llama-swap",
        "responses": {
          "202": {
            "description": "Transition initiated"
          },
          "200": {
            "description": "Already in target state"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Cooldown active — power transition blocked by recent power cycle"
          }
        }
      }
    },
    "/power/off": {
      "post": {
        "summary": "Stop llama-swap and power off the eGPU",
        "responses": {
          "202": {
            "description": "Transition initiated"
          },
          "200": {
            "description": "Already in target state"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Cooldown active — power transition blocked by recent power cycle"
          }
        }
      }
    },
    "/restart": {
      "post": {
        "summary": "Stop then start the complete system",
        "responses": {
          "202": {
            "description": "Transition initiated"
          },
          "200": {
            "description": "Already in target state"
          },
          "409": {
            "description": "Conflict"
          },
          "429": {
            "description": "Cooldown active — power transition blocked by recent power cycle"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "DockMind daemon health (not GPU readiness)",
        "responses": {
          "200": {
            "description": "Daemon is healthy"
          }
        }
      }
    },
    "/containers/{name}/start": {
      "post": {
        "summary": "Start an optional aux container by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container started or already running"
          },
          "404": {
            "description": "Unknown container"
          },
          "409": {
            "description": "Conflict — power transition in progress or state does not allow aux operations"
          },
          "500": {
            "description": "Docker start failed"
          }
        }
      }
    },
    "/containers/{name}/stop": {
      "post": {
        "summary": "Stop an optional aux container by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Container stopped or already stopped"
          },
          "404": {
            "description": "Unknown container"
          },
          "409": {
            "description": "Conflict — power transition in progress or state does not allow aux operations"
          },
          "500": {
            "description": "Docker stop failed"
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Reverse proxy to llama-swap chat completions (OpenAI-compatible)",
        "description": "Proxies the request to the backend AI inference server. Auto-starts the system if it is off. Returns 503 when the backend is unavailable.",
        "responses": {
          "200": {
            "description": "Response from llama-swap"
          },
          "400": {
            "description": "Bad request — forwarded to llama-swap"
          },
          "503": {
            "description": "Service unavailable: backend is off or unreachable. Includes Retry-After header.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the system may retry."
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List available models from llama-swap",
        "description": "Returns cached model list when backend is unavailable. Serves stale cache with X-DockMind-Cached: true header.",
        "responses": {
          "200": {
            "description": "Model list (fresh or cached)",
            "headers": {
              "X-DockMind-Cached": {
                "schema": {
                  "type": "string"
                },
                "description": "Present and set to 'true' when served from cache."
              }
            }
          },
          "503": {
            "description": "Model cache unavailable: no cached data and backend unreachable."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StatusResponse": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": ["Off", "Starting", "Ready", "ShuttingDown", "AwaitingGPUFree", "Error"]
          },
          "gpuPresent": {
            "type": "boolean"
          },
          "gpuName": {
            "type": "string"
          },
          "shellyOn": {
            "type": "boolean"
          },
          "llamaSwapRunning": {
            "type": "boolean"
          },
          "llamaSwapHealthy": {
            "type": "boolean"
          },
          "loadedModels": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Models currently loaded in llama-swap, extracted from the /running endpoint. Empty array when no models are loaded or the service is unhealthy."
          },
          "gpuProcesses": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "pid": { "type": "integer" },
                "name": { "type": "string" }
              }
            },
            "description": "Compute processes currently using the GPU. Empty array when the GPU is absent or no processes are running."
          },
          "lastError": {
            "type": "string",
            "nullable": true
          },
          "cooldownRemaining": {
            "type": "number",
            "description": "Remaining cooldown in seconds. 0 when no cooldown is active."
          },
          "idleRemaining": {
            "type": "number",
            "description": "Remaining seconds before an idle auto-shutdown. 0 when the gateway is disabled, idleTimeout is 0, the state is not Ready, or a request is in flight."
          },
          "auxContainers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "running": { "type": "boolean" }
              }
            },
            "description": "Status of optional aux containers. Empty array when none are configured."
          }
        }
      },
      "OpenAIError": {
        "description": "Error response returned by the OpenAI-compatible gateway when the backend is unavailable.",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Human-readable error message."
              },
              "type": {
                "type": "string",
                "description": "Error type, e.g. 'service_unavailable'."
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code, e.g. 'backend_error', 'startup_timeout', 'model_cache_unavailable'."
              }
            }
          }
        }
      }
    }
  }
}